+++ continual/__init__.py
@@ -1,17 +1,17 @@
-from .batchnorm import BatchNormCo2d  # noqa: F401
-from .conv import ConvCo1d, ConvCo2d, ConvCo3d  # noqa: F401
+from .batchnorm import BatchNorm2d  # noqa: F401
+from .conv import Conv1d, Conv2d, Conv3d  # noqa: F401
 from .delay import Delay  # noqa: F401
 from .interface import _CoModule  # noqa: F401
 from .pooling import (  # noqa: F401
-    AdaptiveAvgPoolCo2d,
-    AdaptiveAvgPoolCo3d,
-    AdaptiveMaxPoolCo2d,
-    AdaptiveMaxPoolCo3d,
-    AvgPoolCo1d,
-    AvgPoolCo2d,
-    AvgPoolCo3d,
-    MaxPoolCo1d,
-    MaxPoolCo2d,
-    MaxPoolCo3d,
+    AdaptiveAvgPool2d,
+    AdaptiveAvgPool3d,
+    AdaptiveMaxPool2d,
+    AdaptiveMaxPool3d,
+    AvgPool1d,
+    AvgPool2d,
+    AvgPool3d,
+    MaxPool1d,
+    MaxPool2d,
+    MaxPool3d,
 )
 from .utils import continual, TensorPlaceholder, Zero  # noqa: F401