+++ CHANGELOG.md
@@ -5,11 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+## Changed
+- Naming to match torch.nn. This lets the continual modules be used as drop-in replacements for torch.nn modules.
 
 ## [0.1.2] - 2021-08-1
 ### Added
-- Pooling modules: MaxPoolCo1d, AvgPoolCo3d, MaxPoolCo3d, AdaptiveAvgPoolCo3d, AdaptiveMaxPoolCo3d
+- Pooling modules: nn.MaxPool1d, nn.AvgPool3d, nn.MaxPool3d, nn.AdaptiveAvgPool3d, nn.AdaptiveMaxPool3d
 
 
 ## [0.1.1] - 2021-08-10